fix: rename attestation entry handler to attestation start handler in message and paired handlers
