+++ src/walletHandlers/attestationStartHandler.js
