+++ src/walletHandlers/pairedHandler.js
@@ -1,13 +1,13 @@
 const eventBus = require('ocore/event_bus.js');
 const mutex = require('ocore/mutex.js');
 
-const attestationEntryHandler = require('./attestationEntryHandler');
+const attestationStartHandler = require('./attestationStartHandler');
 
 eventBus.on('paired', async (device_address, data) => {
     if (data === 'back') return;
 
     const unlock = await mutex.lock(device_address);
-    await attestationEntryHandler(device_address, data);
+    await attestationStartHandler(device_address, data);
 
     return unlock();
 });
